requirements.txt
django
Pillow
torch
torchvision
numpy
scikit-learn
python-decouple
django-crispy-forms
crispy-bootstrap4
celery
redis
psycopg2-binary
setuptools
